App-level libminingbench renaming
