+++ lonestar/mining/cpu/motif-counting/motif.cpp
@@ -1,4 +1,4 @@
-#include "lonestarmine.h"
+#include "MiningBench/Start.h"
 #include "pangolin/BfsMining/vertex_miner.h"
 
 const char* name = "Motif Counting";
