+++ lonestar/mining/cpu/subgraph-listing/sgl_diamond.cpp
@@ -1,4 +1,4 @@
-#include "lonestarmine.h"
+#include "MiningBench/Start.h"
 #include "pangolin/BfsMining/vertex_miner.h"
 
 const char* name = "sgl";
