+++ lonestar/mining/cpu/triangle-counting/tc_mine.cpp
@@ -1,4 +1,4 @@
-#include "lonestarmine.h"
+#include "MiningBench/Start.h"
 #include "pangolin/BfsMining/vertex_miner.h"
 #define TRIANGLE
