+++ lonestar/mining/libminingbench/CMakeLists.txt
@@ -1,4 +1,4 @@
-add_library(miningbench STATIC src/lonestarmine.cpp src/Input.cpp)
+add_library(miningbench STATIC src/Start.cpp src/Input.cpp)
 target_include_directories(miningbench PUBLIC
   "${CMAKE_CURRENT_SOURCE_DIR}/include"
 )
@@ -6,7 +6,7 @@ target_include_directories(miningbench PUBLIC
 target_link_libraries(miningbench Galois::shmem LLVMSupport lonestar)
 
 if(GALOIS_ENABLE_DIST)
-  add_library(miningbench_dist STATIC src/lonestarmine.cpp)
+  add_library(miningbench_dist STATIC src/Start.cpp)
   target_include_directories(miningbench_dist PUBLIC
     "${CMAKE_CURRENT_SOURCE_DIR}/include"
   )
@@ -15,7 +15,7 @@ if(GALOIS_ENABLE_DIST)
 endif()
 
 if(GALOIS_ENABLE_GPU)
-  add_library(miningbench_gpu STATIC src/lonestarmine.cpp src/Input.cpp)
+  add_library(miningbench_gpu STATIC src/Start.cpp src/Input.cpp)
   target_include_directories(miningbench_gpu PUBLIC
     "${CMAKE_CURRENT_SOURCE_DIR}/include"
   )
