+++ lonestar/mining/libminingbench/include/MiningBench/Start.h
