+++ lonestar/mining/libminingbench/src/Input.cpp
@@ -1,4 +1,4 @@
-#include "lonestarmine.h"
+#include "MiningBench/Start.h"
 
 cll::opt<std::string> inputFile(cll::Positional,
                                 cll::desc("<filename: symmetrized graph>"),
